Laptop 11/6
I figured out that the collider issue was all about the triggers that were either on or off on the left barrier and enemy prefab. I need to keep the enemy prefab Is Trigger off in order for the collision to register, but then it will physically collide with the other barriers that I set only to block the player from going out of camera view.