Desktop 11/7
Jeffrey's here! I'm getting close to solving this mystery. The barrier can now detect enemy collision and counts the number of enemiesMissed as well as setting the game to Game Over. However, I have to figure out how to get the layers and triggers set to the right mixture to get exactly what I want.