Laptop 11/12
I fixed the issue where the score wasn't changing when destroying the enemies. I also changed a few small things like the score text in the UI to "Enemies Left" and set the score to decrease when destroying an enemy. However, now the initial score value counted and displayed in the scene doesn't follow the gameManagerScore that I put in the GameManager prefab. I could just change the integer in the GameManager script to a single value, but I want that value to change across levels.