Laptop 11/18
I've finally figured out why the diagonal shooter enemies weren't despawning! It was because the enemies were colliding with the left barrier and couldn't get off-screen, which is why the strike scoring registered while they weren't deleted from the scene. I fixed it by copy and pasting some of the code from the left barrier script to the enemy script and deleting the left barrier script afterwards as I no longer needed it. I've finished Level 2 with changes in the background, swapped the spawn delay values for the diagonal and straight enemy shooters for game balance, and a small change in the UI bar. Finally, I added Level 3 with a new chaser enemy spawner, but I haven't figured out how to make the chaser follow the player while still moving in the same x-direction as the other enemies.